ch4/am: add framework for multiple rndv protocols
A framework that allows multiple RNDV protocols can accommodate special
rndv-like protocols (anything require message matching callbacks) to
work without duplicating code such as the current IPC path.

+++ src/mpid/ch4/src/ch4_types.h
@@ -68,7 +68,7 @@ typedef struct MPIDIG_hdr_t {
     int tag;
     MPIR_Context_id_t context_id;
     int error_bits;
-    uint8_t flags;
+    int flags;
     MPIR_Request *sreq_ptr;
     size_t data_sz;
 } MPIDIG_hdr_t;
